Rename settings.py to config.py
