+++ examples/maplibre_express_circle.py
@@ -1,6 +1,6 @@
 from maplibre import express as mx
+from maplibre.config import settings
 from maplibre.expressions import interpolate
-from maplibre.settings import settings
 
 data = "https://docs.mapbox.com/mapbox-gl-js/assets/earthquakes.geojson"
